Updated index.html, script.js and style.css by moving away from html and using the script.js to get the data from episode.js file
Restructured the index.html and changed the names a bit. We moved on from the using html to create the webpage. Instead converted the html elements to JS by creating variables  in JS and appended the variables inside other variables so that it represents the our html tree.
Also made changes to the pseudocode and css. Added the source in the footer.
To complete level 100, I need to remove the <p> from the episode descriptions and also made the episodeInfoCards fit in the cardsContainer. And probably several other things...